Fix migration 023 SQL types and renumber matview indexes to 024.
Resolves golang-migrate duplicate version conflict and NULL::integer cast for plan_features inserts.

+++ madsan/backend/migrations/023_entitlements_features.up.sql
@@ -14,13 +14,13 @@ INSERT INTO plan_features (plan_id, feature_key, quota_limit)
 SELECT p.id, f.feature_key, f.quota
 FROM plans p
 CROSS JOIN (VALUES
-    ('deal_verification', NULL),
-    ('deal_pack_export', NULL),
-    ('deal_watch', NULL),
-    ('map_premium_layers', NULL),
-    ('supplier_discovery', NULL),
-    ('supplier_portal', NULL),
-    ('api_access', NULL)
+    ('deal_verification', NULL::integer),
+    ('deal_pack_export', NULL::integer),
+    ('deal_watch', NULL::integer),
+    ('map_premium_layers', NULL::integer),
+    ('supplier_discovery', NULL::integer),
+    ('supplier_portal', NULL::integer),
+    ('api_access', NULL::integer)
 ) AS f(feature_key, quota)
 WHERE p.slug IN ('pro', 'enterprise')
 ON CONFLICT DO NOTHING;
